add deleteTopic, refactoring
